Up the iovcnt to a uint32_t
No need to place an artificial limit here. Drivers are
free to set a cap.

Since sendv() is for active messages only and may be limited
to a single MTU, the effective limit is the MTU size (i.e. an
MTU count of 1 byte segments).

+++ include/cci.h
@@ -1549,7 +1549,7 @@ CCI_DECLSPEC int cci_send(cci_connection_t *connection,
 
  */
 CCI_DECLSPEC int cci_sendv(cci_connection_t *connection,
-                           struct iovec *data, uint8_t iovcnt,
+                           struct iovec *data, uint32_t iovcnt,
                            void *context, int flags);
